Day 18 parts 1 and 2
Got stuck on this one for a while, because I misunderstood this part
of the question.

> you must repeatedly do the first action in this list that applies to the snailfish number:

At first, I thought this meant to go through the snail number and look
for the first item that can be exploded or split. But actually you
need to look through the entire snail number for possible explosions,
and only look for splits if there are no pairs to explode.

I realised my mistake after reading this thread on reddit.

https://www.reddit.com/r/adventofcode/comments/rjgfyr/2021_day_18_hint_all_explodes_and_splits_in_the/hp3r4ok/