Day 19 parts 1 and 2
This was the hardest day so far. I got stuck trying to match the
scanner in the test data, and couldn't find the the bug in my
transformations. I ended up implementing the transformations using the
approach here:

https://www.reddit.com/r/adventofcode/comments/rk8qfd/comment/hp8tvir/

That allowed me to solve part 1. I was then able to find the bug in my
original transformations, and implement it the way I originally wanted
to.