Day 24 parts 1 and 2
My initial solution for this took around 30 minutes to
run. @jamienicol pointed out that y is set to zero for every input
digit. This meant that I only had to store the z values instead of (y,
z) pairs, which let me bring the run time down to 20 minutes.

This solution relies on certain properties of the input file
day24.txt. See the assertions in the process_commands method.