Day 23 parts 1 and 2
This was the puzzle that took me the longest to solve this year. My
first attempt at brute solving part 2 didn't find the solution after
checking hundreds of millions of moves for hours.

I managed to solve it after @jamienicol pointed that the amphipods are
not allowed to stop in a space immediately outside a room. I had
overlooked this even though it was the very first extra rule, in bold
type 🤦.